requirements.txt
# Advanced Trading Framework - Requirements
# Core dependencies for the trading system

# Core Python packages
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Data fetching and financial APIs
yfinance>=0.2.18
alpaca-trade-api>=3.0.0

# Machine Learning
scikit-learn>=1.3.0
lightgbm>=4.0.0
joblib>=1.3.0

# Web UI and Visualization
streamlit>=1.28.0
plotly>=5.15.0
matplotlib>=3.7.0

# HTTP and API clients
requests>=2.31.0
aiohttp>=3.8.0
tenacity>=8.2.0

# Logging and utilities
colorama>=0.4.6
psutil>=5.9.0

# Optional: Database support
# sqlalchemy>=2.0.0
# sqlite3  # Built into Python

# Optional: Advanced analytics
# scipy>=1.11.0
# statsmodels>=0.14.0

# Optional: Faster JSON processing
# orjson>=3.9.0

# Development and testing (optional)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0

# Note: Install with pip install -r requirements.txt